--- trailmet/algorithms/quantize/q_utils.py
@@ -1,111 +1,0 @@
-import torch
-import time
-
-class AverageMeter(object):
-    """Computes and stores the average and current value"""
-    def __init__(self, name, fmt=':f'):
-        self.name = name
-        self.fmt = fmt
-        self.reset()
-
-    def reset(self):
-        self.val = 0
-        self.avg = 0
-        self.sum = 0
-        self.count = 0
-
-    def update(self, val, n=1):
-        self.val = val
-        self.sum += val * n
-        self.count += n
-        self.avg = self.sum / self.count
-
-    def __str__(self):
-        fmtstr = '{name} {val' + self.fmt + '} ({avg' + self.fmt + '})'
-        return fmtstr.format(**self.__dict__)
-
-class ProgressMeter(object):
-    """Helper for displaying progress"""
-    def __init__(self, num_batches, meters, prefix=""):
-        self.batch_fmtstr = self._get_batch_fmtstr(num_batches)
-        self.meters = meters
-        self.prefix = prefix
-
-    def display(self, batch):
-        entries = [self.prefix + self.batch_fmtstr.format(batch)]
-        entries += [str(meter) for meter in self.meters]
-        print('\t'.join(entries))
-
-    def _get_batch_fmtstr(self, num_batches):
-        num_digits = len(str(num_batches // 1))
-        fmt = '{:' + str(num_digits) + 'd}'
-        return '[' + fmt + '/' + fmt.format(num_batches) + ']'
-
-
-def accuracy(output, target, topk=(1,)):
-    """Computes the accuracy over the k top predictions for the specified values of k"""
-    with torch.no_grad():
-        maxk = max(topk)
-        batch_size = target.size(0)
-
-        _, pred = output.topk(maxk, 1, True, True)
-        pred = pred.t()
-        correct = pred.eq(target.view(1, -1).expand_as(pred))
-
-        res = []
-        for k in topk:
-            correct_k = correct[:k].reshape(-1).float().sum(0, keepdim=True)
-            res.append(correct_k.mul_(100.0 / batch_size))
-        return res
-
-
-@torch.no_grad()
-def validate_model(val_loader, model, device=None, print_freq=100):
-    if device is None:
-        device = next(model.parameters()).device()
-    else:
-        model.to(device)
-    batch_time = AverageMeter('Time', ':6.3f')
-    top1 = AverageMeter('Acc@1', ':6.2f')
-    top5 = AverageMeter('Acc@5', ':6.2f')
-    progress = ProgressMeter(
-        len(val_loader),
-        [batch_time, top1, top5],
-        prefix='Test: ')
-
-    # switch to evaluate mode
-    model.eval()
-    end = time.time()
-
-    for i, (images, target) in enumerate(val_loader):
-        images = images.to(device)
-        target = target.to(device)
-
-        # compute output
-        output = model(images)
-
-        # measure accuracy and record loss
-        acc1, acc5 = accuracy(output, target, topk=(1,5))
-        top1.update(acc1[0], images.size(0))
-        top5.update(acc5[0], images.size(0))
-
-        # measure elapsed time
-        batch_time.update(time.time()-end)
-        end = time.time()
-
-        if i%print_freq==0:
-            progress.display(i)
-    
-    print(' * Acc@1 {top1.avg:.3f} Acc@5 {top5.avg:.3f}'.format(top1=top1, top5=top5))
-
-    return top1.avg
-
-# To-Do : Modify for balanced subsampling 
-
-def get_train_samples(train_loader, num_samples):
-    train_data = []
-    for batch in train_loader:
-        train_data.append(batch[0])
-        if len(train_data) * batch[0].size(0) >= num_samples:
-            break
-    return torch.cat(train_data, dim=0)[:num_samples]
